Code should not be duplicated - in Annotation/Get.php and Annotation/Post.php
