+++ Annotation/Get.php
@@ -13,120 +13,13 @@
 
 namespace Mmoreram\ControllerExtraBundle\Annotation;
 
-use Mmoreram\ControllerExtraBundle\Annotation\Abstracts\Annotation;
 
 /**
  * Get annotation driver
  *
  * @Annotation
  */
-class Get extends Annotation
+class Get extends RequestBagAnnotation
 {
-    /**
-     * @var string
-     *
-     * The parameter path
-     */
-    public $path;
 
-    /**
-     * @var null|string
-     *
-     * The name that the parameter will get
-     */
-    public $name = null;
-
-    /**
-     * @var mixed
-     *
-     * The default value if the parameter is not set
-     */
-    public $default = null;
-
-    /**
-     * @var bool
-     *
-     * If true, a path like foo[bar] will find deeper items
-     */
-    public $deep = false;
-
-    /**
-     * Gets the path
-     *
-     * @return string
-     */
-    public function getPath()
-    {
-        return $this->path;
-    }
-
-    /**
-     * Sets the path
-     *
-     * @param string $path
-     */
-    public function setPath($path)
-    {
-        $this->path = $path;
-    }
-
-    /**
-     * Gets the name
-     *
-     * @return null|string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * Sets the name
-     *
-     * @param string $name
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-    }
-
-    /**
-     * Gets the default
-     *
-     * @return mixed
-     */
-    public function getDefault()
-    {
-        return $this->default;
-    }
-
-    /**
-     * Sets the default
-     *
-     * @param mixed $default The default value
-     */
-    public function setDefault($default)
-    {
-        $this->default = $default;
-    }
-
-    /**
-     * Is deep
-     *
-     * @return mixed
-     */
-    public function isDeep()
-    {
-        return $this->deep;
-    }
-
-    /**
-     * Sets is deep
-     *
-     * @param boolean $deep
-     */
-    public function setDeep($deep)
-    {
-        $this->deep = $deep;
-    }
 }
